+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/Platform.kt
@@ -1,7 +1,11 @@
 package com.example.bookpedia
 
+import io.ktor.client.engine.HttpClientEngine
+
 interface Platform {
     val name: String
 }
 
 expect fun getPlatform(): Platform
+
+expect fun httpClientEngine(): HttpClientEngine