+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/book/data/SearchResponseDto.kt
@@ -1,0 +1,9 @@
+package com.example.bookpedia.book.data
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SearchResponseDto(
+  @SerialName("docs") val results: List<SearchBookDto>
+)