+++ composeApp/src/commonMain/kotlin/com/example/bookpedia/book/domain/BookRepository.kt
@@ -1,0 +1,8 @@
+package com.example.bookpedia.book.domain
+
+import com.example.bookpedia.core.domain.DataError
+import com.example.bookpedia.core.domain.Result
+
+interface BookRepository {
+  suspend fun searchBooks(query: String): Result<List<Book>, DataError.Remote>
+}