+++ composeApp/src/iosMain/kotlin/com/example/bookpedia/Platform.ios.kt
@@ -1,5 +1,7 @@
 package com.example.bookpedia
 
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.darwin.Darwin
 import platform.UIKit.UIDevice
 
 class IOSPlatform: Platform {
@@ -7,3 +9,5 @@ class IOSPlatform: Platform {
 }
 
 actual fun getPlatform(): Platform = IOSPlatform()
+
+actual fun httpClientEngine(): HttpClientEngine = Darwin.create()