+++ composeApp/src/jvmMain/kotlin/com/example/bookpedia/Platform.jvm.kt
@@ -1,7 +1,12 @@
 package com.example.bookpedia
 
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.okhttp.OkHttp
+
 class JVMPlatform: Platform {
     override val name: String = "Java ${System.getProperty("java.version")}"
 }
 
 actual fun getPlatform(): Platform = JVMPlatform()
+
+actual fun httpClientEngine(): HttpClientEngine = OkHttp.create()